requirements.txt
# Manipulação de documentos Word (.docx)
python-docx==1.1.2

# Dependências do python-docx
lxml>=5.1.0
Pillow>=10.2.0
